scrapyd/requirements.txt
scrapy
scrapy-playwright
html-text
scrapyd
motor